refactor(checker): split module-exports import alias tests
