+++ crates/tsz-checker/tests/conformance_issues/features/mod.rs
@@ -3,5 +3,6 @@ mod elaboration;
 mod function_shape;
 mod implicit_any;
 mod import_aliases;
+mod import_aliases_module_exports;
 mod namespace_construct_signature;
 mod templates;
